requirements.txt
Flask==3.0.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
Werkzeug==3.0.1
Pillow==10.1.0